requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
scikit-learn
matplotlib
seaborn
ipykernel
dill
pyaml
mlflow
# -e .